#19 refactor FAQ formatting
